fix : movements list values
