+++ lib/movements/view/movements_page.dart
@@ -6,6 +6,7 @@ import 'package:warren_task_one/shared/widget/text_page_header.dart';
 import '../../shared/widget/navigation_bar/bottom_nav_bar.dart';
 import '../model/movement_model.dart';
 import '../provider/movement_provider.dart';
+import '../wigdet/movements_empty_body.dart';
 import '../wigdet/movements_list.dart';
 
 class MovementsPage extends HookConsumerWidget {
@@ -50,41 +51,6 @@ class MovementBody extends HookConsumerWidget {
               child: MovementsList(allMovements: allMovements),
             ),
           )
-        ],
-      ),
-    );
-  }
-}
-
-class MovimentsEmptyBody extends StatelessWidget {
-  const MovimentsEmptyBody({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return Center(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Icon(
-            color: const Color.fromARGB(255, 244, 43, 87),
-            Icons.currency_exchange_sharp,
-            size: size.height * .10,
-          ),
-          Padding(
-            padding: EdgeInsets.only(top: size.height * 0.02),
-            child: Text(
-              'Você não possui movimentações',
-              style: TextStyle(
-                fontFamily: "Mansny regular",
-                fontWeight: FontWeight.bold,
-                fontSize: size.height * .025,
-              ),
-            ),
-          ),
         ],
       ),
     );
