+++ lib/movements/wigdet/movements_empty_body.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class MovimentsEmptyBody extends StatelessWidget {
+  const MovimentsEmptyBody({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Center(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(
+            color: const Color.fromARGB(255, 244, 43, 87),
+            Icons.currency_exchange_sharp,
+            size: size.height * .10,
+          ),
+          Padding(
+            padding: EdgeInsets.only(top: size.height * 0.02),
+            child: Text(
+              'Você não possui movimentações',
+              style: TextStyle(
+                fontFamily: "Mansny regular",
+                fontWeight: FontWeight.bold,
+                fontSize: size.height * .025,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
